Create ClassOverrideBook.py
Object-Oriented Library Management System (Python)
This Python code implements a simple library management system using object-oriented programming. It defines the following classes:

Book: This base class represents a book with attributes like title and author. It also includes a __str__ method for printing book details.

EBook: This class inherits from Book and represents an electronic book. It adds an additional attribute size to store the eBook size in megabytes (MB). It overrides the __str__ method to display information specific to eBooks.

PaperBook: This class also inherits from Book and represents a paper book. It adds an attribute num_pages to store the number of pages. It overrides the __str__ method to display information specific to paperbacks.

Library: This class manages a collection of books. It has methods to add books, get the total number of books, and print details of all books in the library.